Create README.md
Developed a real-time object detection algorithm using computer vision and machine learning
techniques. This algorithm can recognize and identify objects in the
user’s environment. Create a user-friendly mobile application that integrates the object
detection system. This application will be the primary interface for blind users to interact
with the system.